金融/数量字段 FloatField → DecimalField 精度修复
问题：所有金融/数量字段用 FloatField（double 精度浮点），多笔成交后
浮点累计误差可达 1.5 万元级别（验证：1000 × 0.1 浮点累加 = 99.9999999999986，
Decimal 累加 = 100.0）。玩家账户不平、撮合分账漂移、财年结算误差。

修复范围（22 个字段）：
- stock: Stock.total_shares/init_net_profit/init_price/current_price、
  StockFundsAccount.cash_balance、StockHolding.shares/cost_price、
  StockOrder.price/quantity/amount、StockCandle.open/high/low/close/change_pct
- fuels: Fuel.price_per_liter
- infrastructures: Infrastructure.price/activation_price
- production_lines: ProductionLine.price/max_per_year
- vehicles: Vehicle.price
- warehouses: Warehouse.capacity/price

保留 FloatField（合理使用，精度足够）：
- Stock.industry_pe/current_carbon/industry_avg_carbon/happiness/pb_random（衍生指标）
- Infrastructure.footprint/employment_rate_bonus/...（加成系数）
- Vehicle.fuel_consumption_per_km/max_cargo/carbon_emission（属性值）
- IndustryField.carbon_emission_coefficient 等（配置项）
- 所有图形坐标 x/y/distance（图形数据）

引擎层改造（apps/stock/engine.py）：
- round2 返回 Decimal 而非 float（消除末尾转 float 丢精度）
- EPS 转 Decimal("1e-9")，撮合循环比较与加减保持 Decimal
- 撮合循环内 cash_map/holding_map/buy_rem/sell_rem/qty/filled 全部 Decimal 累加
- 做市商挂单 base_quantity/base_price/sell_price/buy_price Decimal 化
- compute_match/compute_price/build_candle 保持 float 接口（中间指标），
  调用方显式转 Decimal（DecimalField ORM 读出后）
- 修复 StockCandle 字段映射 bug（build_candle 返回 changePct 驼峰，
  ORM 字段 change_pct 蛇形，原代码 **candle 一直报错）

序列化层（apps/*/serializers.py）：
- 所有金融字段转 DecimalField(max_digits=18, decimal_places=4)，
  min_value 改为 Decimal("0") / Decimal("0.0001")（DRF 警告）
- initPrice/currentPrice read-only 字段同步转 DecimalField

Migrations：Django 自动生成 6 个 0002 alter 迁移文件（无数据回填，
SQLite 上 NUMERIC 与 REAL 底层同类型；切 Postgres 时自动生效）。

验证（端到端）：
1. 1000 × 0.1 累加：Decimal = 100.0（精确），float = 99.9999999999986（修复前）
2. advance_round 端到端：账户/持仓/K 线全部以 Decimal 落库（type=Decimal）
3. manage.py check 通过；migrate 全部应用、零警告
4. vue-tsc 0 错误；vite build 成功（12.74s）
5. DRF DecimalField min_value warning 消除

未变更：前端（DecimalField 序列化为 JSON number，对前端透明）。

数据库：为空（开发环境），本次迁移无数据迁移风险。

+++ backend/apps/fuels/models.py
@@ -6,7 +6,8 @@ class Fuel(models.Model):
     """燃料（比赛级基础数据）。删除比赛时级联删除。"""
 
     name = models.CharField(max_length=255)
-    price_per_liter = models.FloatField()
+    # 燃料单价：玩家购买时累加到现金支出，浮点累计会漂。改 Decimal。
+    price_per_liter = models.DecimalField(max_digits=18, decimal_places=4)
     competition = models.ForeignKey(
         "competitions.Competition",
         on_delete=models.CASCADE,
