+++ backend/apps/production_lines/models.py
@@ -6,9 +6,10 @@ class ProductionLine(models.Model):
     """生产线（比赛级基础数据）。删除比赛时级联删除。"""
 
     name = models.CharField(max_length=255)
-    price = models.FloatField()
+    # 价格 + 年产能：玩家购买场景，必须 Decimal。
+    price = models.DecimalField(max_digits=18, decimal_places=4)
     labor_count = models.IntegerField()
-    max_per_year = models.FloatField()
+    max_per_year = models.DecimalField(max_digits=18, decimal_places=4)
     competition = models.ForeignKey(
         "competitions.Competition",
         on_delete=models.CASCADE,
